cortex: move SortedCortex into synapse.cortex.ram
